requirements.txt
# Web Framework
streamlit>=1.28.0

# GraphQL API
gql[all]>=3.4.0
aiohttp>=3.9.0
graphql-core>=3.2.0

# Image Processing
large-image[all]>=1.25.0
openslide-python>=1.3.0
tifffile>=2023.7.0
scikit-image>=0.21.0
scipy>=1.11.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# ML Models
# PyTorch 2.6.0+ supports Apple Silicon (MPS), CUDA, and CPU
torch==2.6.0
torchvision==0.21.0
# Pin segment-anything to specific commit for reproducibility
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
streamlit>=1.28.0
gql>=3.4.1
requests-toolbelt>=1.0.0
aiohttp>=3.9.0
large-image>=1.24.0
large-image-source-openslide>=1.24.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
torch==2.6.0
torchvision==0.21.0
scikit-image>=0.21.0
matplotlib>=3.7.0
geojson>=3.0.0
shapely>=2.0.0
pandas>=2.0.0
plotly>=5.17.0
pydantic>=2.0.0
